Separate build instructions across UI toolkits
Depending on whether GTK+ 3, Qt4, or Qt5 is used, the build and
runtime dependencies will be different.

The instructions provided specified conflicting package requirements
depending on which "Debian-like" OS is used (check following
paragraphs about reproducibility).

REPRODUCIBILITY

That is a problem for implementing features and causes problems for
packaging the program as a runtime package that works on different
OS'es **even if they are the same OS version**.

I have tried to clarify these requirements according to the UI so that
developing videoplayer-{gtk3,qt} results in a more _reproducible_
development environment.