Remove `sudo` from `apt-get` instructions
Superuser access should not be used by default in _untrusted_ code.  I
am removing `sudo` requests in case a defect somewhere else in the code
breaks the system; the worst case scenario is that this could happen
without the developer even being aware of this.

These instructions are written for developers on Debian-like OS'es.
Every developer on such a system will easily identify such cases as when
`apt-get` is likely to require a `sudo`.